fixup! cmSystemTools: rewrite RemoveRPath using DyanmicEntryList methods

+++ Source/cmSystemTools.cxx
@@ -2541,7 +2541,7 @@ bool cmSystemTools::RemoveRPath(std::string const& file, std::string* emsg,
       elf.GetDynamicEntryPosition(1) - elf.GetDynamicEntryPosition(0);
 
     // Adjust the entry list as necessary to remove the run path
-    int entriesErased = 0;
+    unsigned long entriesErased = 0;
     for (cmELF::DynamicEntryList::iterator it = dentries.begin();
          it != dentries.end();) {
       if (it->first == cmELF::TagRPath || it->first == cmELF::TagRunPath) {
